Address DEBT-442 review feedback
